share clipboard payload detection between the paste paths
`pasted-text.ts` carried its own copy of the clipboard inspection already in
`clipboard-files.ts`. The two paths run back to back on the same event, so any
drift between them means a paste is handled twice or not at all.

Both now import `clipboard-payload.ts`, which owns `browserClipboardFiles`,
`clipboardHasFileEntries`, `clipboardAdvertisesFiles` and `clipboardHasPlainText`,
plus the private type and file-URI helpers. `pasteClipboardFiles` collapses its
separate `advertisesImage` and `advertisesFile` checks into the one call, since
both uses were the disjunction.

The surviving `clipboardHasLocalFileUri` catches per type rather than abandoning
the scan on the first `getData` that throws, so a readable `text/uri-list` is no
longer hidden by an unreadable sibling type.

Relative imports carry the `.ts` extension: `npm test` runs
`node --experimental-strip-types`, whose ESM resolver does not add it.
