+++ studio/frontend/src/features/chat/utils/pasted-text.ts
@@ -4,6 +4,11 @@
 // Long pastes become a .txt attachment instead of flooding the composer.
 // TextAttachmentAdapter still sends the full text to the model.
 
+import {
+  clipboardAdvertisesFiles,
+  clipboardHasFileEntries,
+} from "./clipboard-payload.ts";
+
 const PASTED_TEXT_MIME = "text/plain";
 export const PASTED_TEXT_MIN_CHARS = 2000;
 export const PASTED_TEXT_MIN_LINES = 40;
@@ -203,42 +208,6 @@ export function attachmentsSample(
   return "";
 }
 
-function clipboardHasFiles(clipboardData: DataTransfer): boolean {
-  if (Array.from(clipboardData.files).some((file) => file.size > 0))
-    return true;
-  return Array.from(clipboardData.items).some((item) => item.kind === "file");
-}
-
-function clipboardHasLocalFileUri(
-  clipboardData: DataTransfer,
-  types: readonly string[],
-): boolean {
-  return types
-    .filter((type) => type.includes("uri-list") || type.includes("urilist"))
-    .some((type) => {
-      try {
-        return clipboardData
-          .getData(type)
-          .split(/\r?\n/)
-          .some((line) => line.trim().toLowerCase().startsWith("file:"));
-      } catch {
-        return false;
-      }
-    });
-}
-
-// Native (Tauri) images and copied files are advertised by type only, with no
-// entry in files or items, and pasteClipboardFiles reads them itself.
-function clipboardAdvertisesFiles(clipboardData: DataTransfer): boolean {
-  const types = Array.from(clipboardData.types, (type) => type.toLowerCase());
-  return (
-    types.some((type) => type.startsWith("image/")) ||
-    types.includes("files") ||
-    types.some((type) => type.includes("copied-files")) ||
-    clipboardHasLocalFileUri(clipboardData, types)
-  );
-}
-
 function clipboardText(clipboardData: DataTransfer): string {
   try {
     return clipboardData.getData("text/plain");
@@ -257,7 +226,7 @@ export function pasteLongTextAsFile(
   const { clipboardData } = event;
   if (!clipboardData) return false;
   // Images and files keep the existing paste path.
-  if (clipboardHasFiles(clipboardData)) return false;
+  if (clipboardHasFileEntries(clipboardData)) return false;
   if (clipboardAdvertisesFiles(clipboardData)) return false;
 
   const text = clipboardText(clipboardData);
